flash: reflect updates to flash driver in platform
change the read length for spi read from 16b to 32b for longer reads

+++ software/inc/flash.h
@@ -7,6 +7,6 @@
 */
 void flash_eraseSector(uint32_t addr);
 void flash_programPage(uint32_t addr, uint8_t* tx, uint8_t* rx, uint16_t len);
-void flash_SPIRead(uint32_t addr, uint8_t* tx, uint8_t* rx, uint16_t len);
+void flash_SPIRead(uint32_t addr, uint8_t* tx, uint8_t* rx, uint32_t len);
 
 #endif
